Move files around
Also included import changes to avoid wrong import underlining in Pycharm IDE
