--- AlgorithmAnalyzer/Backend/convert_audio_wrapper/__init__.py
@@ -1,4 +1,0 @@
-# instructions to be run when convert_audio_wrapper is imported
-
-from .convert_webm import *
-
