+++ AlgorithmAnalyzer/Backend/utils/__init__.py
@@ -1,3 +1,3 @@
-from .SampleManager import SampleManager, UsernameException
+from Backend.utils.sample_manager.SampleManager import SampleManager, UsernameException
 
 __all__ = [SampleManager, UsernameException]
