+++ AlgorithmAnalyzer/Backend/utils/speech_recognition_wrapper/tests/tests_speech_recognition.py
@@ -3,8 +3,7 @@ import sys
 import unittest
 import speech_recognition as sr
 
-from speech_recognition_wrapper import speech_to_text_wrapper as wrapper
-
+from Backend.utils.speech_recognition_wrapper import speech_to_text_wrapper as wrapper
 
 
 class TestSpeechToText(unittest.TestCase):
